+++ R/AAAA.R
@@ -1,22 +1,26 @@
 #' @importFrom assertthat validate_that assert_that
-#' @importFrom tidyselect all_of eval_select eval_rename contains
+#' @importFrom tidyselect all_of eval_select eval_rename contains everything
 #' @importFrom dplyr mutate bind_rows select ungroup summarise left_join
 #' @importFrom dplyr group_modify rename count vars group_by n first last
 #' @importFrom dplyr case_when filter arrange group_vars distinct bind_cols
-#' @importFrom dplyr groups slice everything mutate_at across
+#' @importFrom dplyr groups slice mutate_at across
 #' @importFrom purrr map_dfr walk partial map map_chr modify flatten_chr imap
 #' @importFrom purrr flatten_int map_lgl modify_if map_int map2 keep flatten
+#' @importFrom purrr list_flatten
 #' @importFrom tidyr pivot_wider pivot_longer replace_na fill separate unite
 #' @importFrom tidyr complete
 #' @importFrom forcats fct_inorder
 #' @importFrom rlang sym syms quo_get_expr as_string := .data .env is_empty
-#' @importFrom rlang enquo enquos is_named is_atomic flatten_if have_name
+#' @importFrom rlang enquo enquos is_named is_atomic have_name
+#' @importFrom rlang env_clone
 #' @importFrom glue glue
 #' @importFrom tibble tibble as_tibble is_tibble
-#' @importFrom stats median rnorm sd na.omit setNames
+#' @importFrom stats median rnorm sd na.omit setNames update
 #' @importFrom utils capture.output packageVersion str
 #' @importFrom stringr fixed str_split str_count str_detect str_replace
+#' @importFrom stringr str_extract
 #' @importFrom tools file_ext
+#' @importFrom lifecycle deprecate_warn
 #'
 #' @include summary-functions.R
 #' @include utils.R